Factor out SimpleTestCircuit
