+++ transactron/testing/__init__.py
@@ -2,6 +2,7 @@ from amaranth.sim._async import TestbenchContext, ProcessContext, SimulatorConte
 from .input_generation import *  # noqa: F401
 from .functions import *  # noqa: F401
 from .infrastructure import *  # noqa: F401
+from .test_circuit import *  # noqa: F401
 from .method_mock import *  # noqa: F401
 from .testbenchio import *  # noqa: F401
 from .profiler import *  # noqa: F401
